The message should say "less than or equal <number>"
The validator is checking whether passed number is greater than an @option,
If this is true then the exception is raised. However when you pass number equal
the @option then exception is not raised. To make the message correct 
it should display "less than _or equal_ <number>" instead of just "less than".